Combine useSite and useSites
