+++ src/common/dataAdaptors/index.js
@@ -1,1 +1,2 @@
+export * from './siteAdaptors'
 export * from './widgetAdaptors'
