+++ src/components/Languages/LanguagesData.js
@@ -1,6 +1,6 @@
 // FPCC
-import useSites from 'common/dataHooks/useSites'
-import useMySites from 'common/dataHooks/useMySites'
+import { useSites } from 'common/dataHooks/useSites'
+import { useMySites } from 'common/dataHooks/useMySites'
 import { parentLanguageColors } from 'assets/parentLanguageColors'
 
 function LanguagesData() {
